requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-decouple>=3.8
Pillow>=10.0.0
django-filter>=23.0
dj-database-url>=2.0.0
psycopg2==2.8.6
